Added Role based access control

+++ app/posts/routes.py
@@ -1,5 +1,6 @@
 from flask import Blueprint, request, jsonify
 from flask_jwt_extended import jwt_required, get_jwt_identity
+from app.auth.utils import admin_required
 
 from app.extensions import db, jwt
 from app.posts.models import Post
@@ -11,7 +12,7 @@ post_bp = Blueprint("posts", __name__)
 
 @post_bp.post("/posts")
 @jwt_required()  # verifies that the endpoint cannot be acces without a token and is valid token compares with the token that is sent to the client at the time of login 
-def create_post(user_id):
+def create_post():
 
     current_user_id = get_jwt_identity()  # extracts the identity like id or username from the token
 
@@ -25,6 +26,8 @@ def create_post(user_id):
     db.session.commit()
 
     return jsonify(post_schema.dump(post)), 201
+
+
 
 
 @post_bp.get("/users/<int:user_id>/posts")
@@ -50,10 +53,18 @@ def update_post(post_id):
     return jsonify(post_schema.dump(post)), 200 
 
 
+# first check - Is this post owned by the logged-in user?
+
 @post_bp.delete("/posts/<int:post_id>")
+@jwt_required()
 def delete_post(post_id):
 
+    current_user_id  = get_jwt_identity()
+
     post = Post.query.get_or_404(post_id)
+
+    if post.user_id != current_user_id:
+         return jsonify({"error": "You cannot delete this post"}), 403 # 403 - logged in but now allowed
 
     db.session.delete(post)
     db.session.commit()
